refac(web): module splitting
