+++ Cargo.toml
@@ -21,3 +21,5 @@ tokio = { version = "0.2", features = ["macros"]}
 mockito = "0.25"
 assert-impl = "*"
 env_logger = "0.7"
+serde_json = { version = "1.0", features = ["preserve_order"] }
+
