+++ rustfmt.toml
@@ -1,0 +1,1 @@
+format_code_in_doc_comments = true
