+++ src/web_api/response/mod.rs
@@ -2,7 +2,9 @@ use crate::web_api::domain::LoginInfos;
 use crate::web_api::domain::UserFeed;
 use crate::web_api::domain::UserInfos;
 
-use serde::{Deserialize, Deserializer};
+use serde::Deserialize;
+
+mod deserializer;
 
 impl std::convert::From<reqwest::Error> for UserInfosError {
     fn from(_error: reqwest::Error) -> Self {
@@ -29,7 +31,7 @@ pub struct GraphQLResponse<T> {
 
 #[derive(Debug, Deserialize)]
 pub struct UserFeedResponse {
-    #[serde(rename = "user", deserialize_with = "nested_user_feed")]
+    #[serde(rename = "user", deserialize_with = "deserializer::nested_user_feed")]
     pub feed: UserFeed,
 }
 
@@ -74,16 +76,4 @@ pub struct PhoneVerificationSettings {
     pub resend_sms_delay_sec: i32,
     pub robocall_count_down_time_sec: i32,
     pub robocall_after_max_sms: bool,
-}
-
-fn nested_user_feed<'de, D>(deserializer: D) -> Result<UserFeed, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    #[derive(Deserialize)]
-    struct A {
-        edge_owner_to_timeline_media: UserFeed,
-    }
-
-    A::deserialize(deserializer).map(|a| a.edge_owner_to_timeline_media)
 }
