+++ tests/web_api_client/authenticated_client.rs
@@ -9,7 +9,7 @@ fn behaviour() {
     assert_impl!(FetchUserInfos: Client);
     assert_impl!(!FetchUserFeed: Client);
     // TODO implementation needed
-    assert_impl!(!FetchMediaInfo: Client);
+    assert_impl!(!FetchMediaInfos: Client);
     assert_impl!(!FetchMediaComments: Client);
     assert_impl!(!Search: Client);
     assert_impl!(!FetchTagFeed: Client);
