+++ tests/web_api_client/unauthenticated_client.rs
@@ -10,7 +10,7 @@ fn behaviour() {
     assert_impl!(FetchUserFeed: Client);
 
     // TODO implementation needed
-    assert_impl!(!FetchMediaInfo: Client);
+    assert_impl!(!FetchMediaInfos: Client);
     assert_impl!(!FetchMediaComments: Client);
     assert_impl!(!Search: Client);
     assert_impl!(!FetchTagFeed: Client);
